dialog view and window merged
