--- src/NexusMods.App.UI/Dialog/DialogView.axaml
@@ -1,116 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:vm="clr-namespace:NexusMods.App.UI.Dialog"
-             xmlns:controls="clr-namespace:NexusMods.App.UI.Controls"
-             xmlns:panels="clr-namespace:Avalonia.Labs.Panels;assembly=Avalonia.Labs.Panels"
-             xmlns:reactiveUi="http://reactiveui.net"
-             xmlns:icons="clr-namespace:NexusMods.UI.Sdk.Icons;assembly=NexusMods.UI.Sdk"
-             mc:Ignorable="d" d:DesignWidth="400" 
-             x:Class="NexusMods.App.UI.Dialog.DialogView"
-             d:DataContext="{x:Static vm:DialogDesignViewModel.StandardContent}">
-
-    <!-- Temporary styles to see modal button variations -->
-    <UserControl.Styles>
-        <Style Selector="controls|StandardButton">
-            <Style Selector="^ /template/ TextBlock#PART_Label">
-                <Setter Property="Theme" Value="{StaticResource BodySMSemiTheme}" />
-            </Style>
-
-            <Style Selector="^.Default">
-                <Setter Property="Background" Value="{StaticResource NeutralStrongBrush}" />
-                <Setter Property="Foreground" Value="{StaticResource NeutralInvertedBrush}" />
-
-                <Style Selector="^:pointerover">
-                    <Setter Property="Background" Value="{StaticResource NeutralSubduedBrush}" />
-                </Style>
-            </Style>
-
-            <Style Selector="^.Danger">
-                <Setter Property="Background" Value="{StaticResource DangerStrongBrush}" />
-                <Setter Property="Foreground" Value="{StaticResource NeutralInvertedBrush}" />
-
-                <Style Selector="^:pointerover">
-                    <Setter Property="Background" Value="{StaticResource DangerSubduedBrush}" />
-                </Style>
-            </Style>
-
-            <Style Selector="^.Info">
-                <Setter Property="Background" Value="{StaticResource InfoStrongBrush}" />
-                <Setter Property="Foreground" Value="{StaticResource NeutralInvertedBrush}" />
-
-                <Style Selector="^:pointerover">
-                    <Setter Property="Background" Value="{StaticResource InfoSubduedBrush}" />
-                </Style>
-            </Style>
-
-            <Style Selector="^.Premium">
-                <Setter Property="Background" Value="{StaticResource PremiumStrongBrush}" />
-                <Setter Property="Foreground" Value="{StaticResource NeutralInvertedBrush}" />
-
-                <Style Selector="^:pointerover">
-                    <Setter Property="Background" Value="{StaticResource PremiumSubduedBrush}" />
-                </Style>
-            </Style>
-
-            <Style Selector="^.Primary">
-                <Setter Property="Background" Value="{StaticResource PrimaryStrongBrush}" />
-                <Setter Property="Foreground" Value="{StaticResource NeutralInvertedBrush}" />
-                <Style Selector="^:pointerover">
-                    <Setter Property="Background" Value="{StaticResource PrimarySubduedBrush}" />
-                </Style>
-                <Style Selector="^:pressed">
-                    <Setter Property="Background" Value="{StaticResource PrimaryWeakBrush}" />
-                </Style>
-            </Style>
-        </Style>
-    </UserControl.Styles>
-
-    <Grid RowDefinitions="Auto,*">
-
-        <!-- titlebar -->
-        <Grid x:Name="Titlebar" Grid.Row="0"
-              ColumnDefinitions="*,Auto"
-              Margin="16,16,16,0">
-
-            <!-- has to have IsHitTestVisible="False" so the mouse events are passed through to the OS for 
-                custom titlebar dragging -->
-            <TextBlock Grid.Column="0" x:Name="TitleTextBlock"
-                       Text="{Binding WindowTitle}"
-                       Foreground="{StaticResource NeutralStrongBrush}"
-                       Theme="{StaticResource BodySMSemiTheme}"
-                       VerticalAlignment="Center"
-                       ToolTip.Tip="Close"
-                       IsHitTestVisible="False"
-                       TextWrapping="NoWrap"
-                       TextTrimming="CharacterEllipsis" />
-
-            <controls:StandardButton Grid.Column="1" x:Name="CloseButton"
-                                     DockPanel.Dock="Right"
-                                     LeftIcon="{x:Static icons:IconValues.Close}"
-                                     ShowIcon="IconOnly"
-                                     ShowLabel="False"
-                                     Fill="None"
-                                     Type="Tertiary"
-                                     VerticalAlignment="Center" />
-        </Grid>
-
-        <!-- content -->
-        <Border x:Name="ContentBorder" Grid.Row="1">
-
-            <!-- content, buttons -->
-            <Grid RowDefinitions="*, Auto" Margin="16,16,16,16">
-
-                <Border Grid.Row="0" x:Name="ContentContainer">
-                    <reactiveUi:ViewModelViewHost x:Name="ContentViewModelHost" />
-                </Border>
-
-                <!-- buttons -->
-                <panels:FlexPanel x:Name="ButtonsFlexPanel" Grid.Row="1"
-                                  ColumnSpacing="8" Margin="0 16 0 0" />
-            </Grid>
-
-        </Border>
-    </Grid>
-</UserControl>
