+++ src/NexusMods.App.UI/NexusMods.App.UI.csproj
@@ -752,6 +752,9 @@
         <Compile Update="Dialog\DialogViewModel.cs">
           <DependentUpon>IDialogViewModel.cs</DependentUpon>
         </Compile>
+        <Compile Update="Dialog\DialogDesignViewModel.cs">
+          <DependentUpon>IDialogViewModel.cs</DependentUpon>
+        </Compile>
     </ItemGroup>
 
     <ItemGroup>
